# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Service Marketplace API
  description: |
    REST API for Service Marketplace - a platform connecting service providers with customers.

    ## Authentication
    - Use `/api/auth/register` to create an account
    - Use `/api/auth/login` to get JWT tokens
    - Include `Authorization: Bearer <token>` header for authenticated endpoints

    ## Roles
    - **CUSTOMER**: Browse services, create bookings, leave reviews
    - **VENDOR**: Manage services, view bookings, track earnings
    - **ADMIN**: Full system access
  contact:
    name: Hien Nguyen
    email: fangrixian.nguyenn@outlook.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 0.0.1
servers:
- url: http://localhost
  description: Generated server url
security:
- bearerAuth: []
...
paths:
...
  /api/reviews:
    post:
      tags:
      - Reviews
      summary: Create review
      description: Leave a review for a booking the caller owns. The booking must
        be COMPLETED and not already reviewed.
      operationId: createReview
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewCreateRequest"
        required: true
      responses:
        "201":
          description: Review created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewResponse"
        "400":
          description: Invalid input
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewResponse"
        "404":
          description: Booking not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewResponse"
        "422":
          description: "Not your booking, not completed, or already reviewed"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewResponse"
      security:
      - bearerAuth: []
...
  /api/auth/register:
    post:
      tags:
      - Authentication
      summary: Register new user
      description: Create a new user account. Optionally register as provider.
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
        required: true
      responses:
        "201":
          description: User registered successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already registered
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input data
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/auth/refresh:
    post:
      tags:
      - Authentication
      summary: Refresh access token
      description: Get a new access token using a valid refresh token
      operationId: refreshToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
        required: true
      responses:
        "200":
          description: Token refreshed successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TokenRefreshResponse"
        "401":
          description: Invalid or expired refresh token
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TokenRefreshResponse"
  /api/auth/login:
    post:
      tags:
      - Authentication
      summary: Login
      description: Authenticate with email and password to get JWT tokens
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: Login successful
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input data
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /api/reviews/provider/{providerId}:
    get:
      tags:
      - Reviews
      summary: Get provider reviews
      description: "List all reviews for a provider, newest first."
      operationId: getProviderReviews
      parameters:
      - name: providerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Reviews retrieved
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewResponse"
  /api/reviews/service/{serviceId}:
    get:
      tags:
      - Reviews
      summary: Get service reviews
      description: "List all reviews for a service, newest first."
      operationId: getServiceReviews
      parameters:
      - name: serviceId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Reviews retrieved
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewResponse"
...
  /api/auth/me:
    get:
      tags:
      - Authentication
      summary: Get current user
      description: Return the authenticated user's profile from the JWT.
      operationId: getCurrentUser
      responses:
        "200":
          description: Current user profile
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Not authenticated
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    ReviewCreateRequest:
      type: object
      properties:
        bookingId:
          type: integer
          format: int64
        rating:
          type: integer
          format: int32
          maximum: 5
          minimum: 1
        comment:
          type: string
          maxLength: 2000
          minLength: 0
      required:
      - bookingId
      - rating
    ReviewResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        bookingId:
          type: integer
          format: int64
        serviceId:
          type: integer
          format: int64
        serviceTitle:
          type: string
        customerId:
          type: integer
          format: int64
        customerName:
          type: string
        providerId:
          type: integer
          format: int64
        providerName:
          type: string
        rating:
          type: integer
          format: int32
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
...
    RegisterRequest:
      type: object
      properties:
        fullName:
          type: string
          maxLength: 100
          minLength: 2
        email:
          type: string
          minLength: 1
        password:
          type: string
          maxLength: 100
          minLength: 8
          pattern: "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).+$"
        phoneNumber:
          type: string
          minLength: 1
          pattern: "^\\+?[1-9]\\d{1,14}$"
        registerAsProvider:
          type: boolean
      required:
      - email
      - fullName
      - password
      - phoneNumber
    AuthResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
        accessToken:
          type: string
        refreshToken:
          type: string
        accessTokenExpiresAt:
          type: string
          format: date-time
        refreshTokenExpiresAt:
          type: string
          format: date-time
    RefreshTokenRequest:
      type: object
      properties:
        refreshToken:
          type: string
          minLength: 1
      required:
      - refreshToken
    TokenRefreshResponse:
      type: object
      properties:
        accessToken:
          type: string
        accessTokenExpiresAt:
          type: string
          format: date-time
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
      required:
      - email
      - password
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fullName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        role:
          type: string
          enum:
          - CUSTOMER
          - VENDOR
          - ADMIN
        status:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          - SUSPENDED
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      description: Enter JWT token (without 'Bearer ' prefix)
      scheme: bearer
      bearerFormat: JWT